requirements.txt
numpy
yfinance
pandas
streamlit
requests
python-dotenv
python-dateutil
matplotlib